+++ dpg_navigator/__init__.py
@@ -32,7 +32,7 @@ from ._preview import (
 from ._pdf import pdf_available
 from ._html import html_available
 
-__version__ = "1.0.0b2"
+__version__ = "1.0.0b3"
 
 __all__ = [
     "FileDialog",